+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.11
 
+* Upfill `Nav` and `UIKitNav`
+
 ## 0.0.10
 
 * Filling of `UIKitUtility`